dpaa_eth: Add overhead accounting control to CGRs
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

QMan can account for a fixed overhead when comparing the current byte
counts against the CGR notification threshold. We can use it to better
memory management in those cases when the fixed overhead from skb shells
and the netdev's needed_headroom is significant compared to the outgoing
frame size.
Unfortunately, with the current hardware the maximum value of
the overhead accounting limit (OAL) is 127, which is not enough for our
needs; but even so, it is better than no OAL at all.

Change-Id: I368a59816b52b176aa569eb446a869ac20b3d8b3
Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
(cherry picked from commit 4084d9c8d258aad9a1fb3367fd8c81f01a9c9e30)
Reviewed-on: http://git.am.freescale.net:8181/1039
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>