dpa_offload: Add packet classification component
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The packet classification offloading driver implements its functionalities
using the NetCommSw FMan driver. It exposes an easy-to-use API which can be
called either from kernel space or from user space (through the
classification offloading driver wrapper).

It is able to create or import (from FMD resources) 3 types of tables -
exact match table, indexed table and HASH table. Imported tables can be
either empty or prefilled.

It offers an API to insert, remove, or modify entries. It allows the users
to classify and
     - enqueue
     - multicast
     - discard
     - re-classify or
     - return-to-KeyGen
network packets.

It is able to create or import (from FMD resources) header manipulation
operations and attach them to table entries. It allows runtime modification
of existing (created or imported) header manipulation operations.

It offers an API to create or import (from FMD resources) multicast groups.
It allows the user to add or remove members to existing multicast groups.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Change-Id: I854c0c3c2eba6d6f441cb46e502e6dbc623c48d5
Reviewed-on: http://git.am.freescale.net:8181/2233
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>