dpa_offload: Fix for dpa_stats counter selection
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The selection of a counter that has offset zero caused returning of
default counter value(zero) for all selections matching this patters.
The reassembly NCSP and all the Classif CC Node counters supported on
FMAN v3 are now marked with UNSUPPORTED_CNT_SEL and only for them the
default counter value is returned.

Change-Id: I52cdca2cb65765af88b8e43517d66c88626a7143
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2669
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Floarea Anca Jeanina-B12569 <anca.floarea@freescale.com>
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>