dpaa_eth: Fix pool counters on error path
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

When system memory is depleted, the Rx buffers are put back into
their pool. Bad accounting on that error path made it so the software
software counters remained behind the hardware counters, eventually
allowing for negative totals.
Fix this and keep the software pool counters in sync with the hardware
counters.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I46191bc38f12aa6a25378eba3e5891bc3f18cbc6
Reviewed-on: http://git.am.freescale.net:8181/3049
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>