dpa_offload: Add separate resource pointer for multicast group import operation
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

When multicast group was preallocated from a xml PCD file, the handle to the group
was given as a multicast group parameter.
To align the multicast component with header manipulation and classifier
components, the group handle will be given as a resource input parameter
to the function that creates a multicast group.

Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Change-Id: I254d1c681cc016513d4cb569e0dfcf5c56478163
Reviewed-on: http://git.am.freescale.net:8181/3279
Reviewed-by: Chereji Marian-Cornel-R27762 <marian.chereji@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>