dpaa_eth: Move bpool seeding back to probe for SG driver
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

Only the non-SG ("optimized for forwarding") flavour of the driver
needs to have the buffer pool seeding done after all interfaces are
probed, as in this case the buffer size may vary between interfaces.

For the termination optimized flavour, we have a fixed buffer size on Rx,
so there's no need to wait for the pool seeding.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I101ffa38e9f83f90a432620898a2a24ede6c0895
Reviewed-on: http://git.am.freescale.net:8181/3656
Reviewed-by: Hamciuc Bogdan-BHAMCIU1 <bogdan.hamciuc@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>