qbman: Clarify post-condition for bman_acquire()
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

If bman_acquire() cannot provide the requested number of buffers, it
doesn't touch the bpool, even if that may still contain some, fewer,
buffers. -ENOMEM is returned, but the @bufs array is overwritten
nevertheless.

Make it clear that the caller should not rely on @bufs if this was the
case.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Change-Id: I5b4336fb98bbaa67924600ba5bad9c9e2d665509
Reviewed-on: http://git.am.freescale.net:8181/3646
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>