dpa_offload: Allow 0xff masks for HASH classifier table entries
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

As key masks are not supported in HASH tables, the DPA Classifier was
refusing to add an entry to a HASH table with a key that had
any mask (mask pointer not NULL). This was changed so that the DPA
Classifier can recognize a 0xff mask all over and ignore it. Any other
(more complex) mask will still be refused.

Change-Id: I0732d873fa5d10def4f2646c2bd248aadcffa736
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4911
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>