Prevent kernel panic if USDPAA mem cannot be allocated
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

USDPAA DMA memory is allocated very early in the boot process
using the memblock_alloc functions.  The old function would
call panic() in cases that the allocation would fail, but
no indication would be given on the console since the
drivers were not loaded yet.  This patch changes the function
called so that the panic will not occur and debug info can be
seen in the boot log

Signed-off-by: Roy Pledge <Roy.Pledge@freescale.com>
Change-Id: I8ebe9eb826973ea8ef6d16ba6fc73512b779ab77
Reviewed-on: http://git.am.freescale.net:8181/5953
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Haiying Wang <Haiying.Wang@freescale.com>
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>