drm: xilinx: plane: Propagate the manager's format when no plane format is found
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

Propate the manager's format to the plane when no plane format is found.
This happens when there's no format conversion IPs in the pipeline,
and the plane needs to be set to the manager's format.

Signed-off-by: Hyun Kwon <hyunk@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit eedf80b038937e01a70650e35347faf34100e084)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>