fsl_qman: some fixes in ceetm ccg APIs
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

In ceetm_ccg_set(), check whether the cb_node is already in the list
before adding it to the list in order to make sure the cb_node will
only be added to list once. Also update the CSCN target only when
CSCN_EN mask is set. In ceetm_ccg_release(), check whether the
ccg->cb_node is in the list before deleting it.

Signed-off-by: Haiying Wang <Haiying.Wang@freescale.com>
Change-Id: Ie48e3459d0c856d91fa84ff74676d2bbe69850ae
Reviewed-on: http://git.am.freescale.net:8181/11999
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>