dpa_offload: Fix DPA Stats async request for 64bit US mode
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

Because in the case of 64 bit user-space, 64 bit kernel-space
CONFIG_COMPAT is enabled, the DPA Stats asynchronous request failed to
work. Solving the problem involves analyzing at run time the environment
and taking the decision if the compat or normal function for read is
used.
Also some changes in the internal DPA Stats data structures are needed.

Change-Id: I0d7357def19e42901a3c9b90478b07ef53f62301
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/12052
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Tested-by: Jose Rivera <German.Rivera@freescale.com>