tdm/test: Added SLIC loopback test.
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

Modified test to cover SLIC loopback also. This test can
now be run for internal as well as external loopback. To
choose which test (internal/external) is to be executed
can be controlled by module param "in_loopback". When
in_loopback=1 test is configured for internal loopback,
when in_loopback=0 test is configured for external(SLIC)
loopback.

Signed-off-by: Sandeep Singh <sandeep@freescale.com>
Change-Id: I1244e9930e00ea7e17cff28f28e351814cc0d249
Reviewed-on: http://git.am.freescale.net:8181/12414
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Poonam Aggrwal <poonam.aggrwal@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>