dpaa_eth: Delay buffer pool seeding for private ports
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

Until now, all private ports had to be configured to use the same
buffer size, as they all share a common buffer pool.

Create the framework for working with different buffer sizes on the
private ports; a single, common buffer pool is still used by all
private ports.
The pool is seeded with buffers large enough to accomodate frames
for all private ports. In order to be able to compute the largest
needed buffer size, the buffer pool seeding is delayed until the
first private interface is raised.

Temporarily disable the boot-time ethernet unit test, which requires
the use of the common buffer pool.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I22d40d136e35320b7b4af76e18e53faff7460cbf
Reviewed-on: http://git.am.freescale.net:8181/1466
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Signed-off-by: Lu.Jiang <lu.jiang@windriver.com>