mac: make the dt parser endianness-free
The OF API used in the mac dev driver parses the device tree and uses raw
pointers to data. If the CPU is LE the values read will be mixed and the device
will not probe.

This patch belongs to the big effort of porting eth & mac drivers to ARM
architecture.

Signed-off-by: Rotariu Marian Cristian-B44860 <marian.rotariu@freescale.com>
Change-Id: Ibb7893d950e92bf4038cf565172a77432adc1055
Reviewed-on: http://git.am.freescale.net:8181/20928
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/32498
[Yang: Original patch taken from
QorIQ-SDK-V1.8-SOURCE-20150619-yocto.iso]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>