dpaa_eth: convert fq dts entries to CPU endianess
The data related to fqs from dts are in BE. In order to be used by the CPU, the
data need to be converted into CPU endianess. Because the current fqids
structure is used in many ways and in many files, the code for endianess
conversion is not localized. To make this patch as simple as possible, an
intermediary variable was used.

Change-Id: I1beec33fa3ce816899e3bb02e0cf88ed0b0d0520
Signed-off-by: Rotariu Marian Cristian-B44860 <marian.rotariu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/20964
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/32495
[Yang: Original patch taken from
QorIQ-SDK-V1.8-SOURCE-20150619-yocto.iso]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>