gianfar: Disable Rx snooping for LS1
It was observed that Rx snooping incurs a significant
performance penalty.  My suspicion is that it interferes
with the DMA software coherency, currently performed
by DMA-API.  It is known that there are H/W coherency
issues at least for the Tx side (see read data interleaving
issue) which is why the software coherency is needed in the
first place.  This patch disables Rx snooping as well, as a
workaround, until the root cause for the Rx side coherency
related issue is found.

Change-Id: I00c3a5723d52d2f9a9d18e11b5c64c9c208c466b
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/25325
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>
Reviewed-by: Matthew Weigel <Matthew.Weigel@freescale.com>
[Xulin: Original patch taken from
Freescale-Linux-SDK-for-LS1021A-IOT-Rev2-v0.4-SOURCE-20150907-yocto.iso]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>