dpaa_eth: add VSP compatibility for eth ports
This patch provides compatibility for the Ethernet driver ports with the VSP
capability. Until now, the FMan ports managed by the Ethernet driver were not
able to properly configure VSP because of the buffer layout of the FMan ports
used.

When VSPs are not enabled, the RX and TX path buffer layout are different, the
Ethernet driver maintains separate buffer layouts for transmission and
reception. When VSPs are enabled on a certain FMan port, the buffer layouts for
Rx and Tx should be the same, the same buffer layout being used inside a
storage profile.

In order to bypass this the VSP capability is disabled on the Tx path by setting
ContextB to 0 in the FQ descriptor of the queues that are transmitting the
frames to the Tx port.

Change-Id: Ie29b8a0fe8b0b9c2344658f8e4a35e0392b6bf8e
Signed-off-by: Marian-Cristian Rotariu <marian.rotariu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/25613
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
[Yang: Original patch taken from
QorIQ-SDK-V1.8-SOURCE-20150619-yocto.iso]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>