dpaa_eth: use proper recycling condition
The condition that needs to be evaluated to determine if a certain
fd is to be recycled (buffers released in the buffer pool by HW)
is (fd.bpid != 0xff). Several places in the code were using the
previous condition (fd.cmd & FM_FD_CMD_FCO) that is always true
after the unification of the recycling and confirmation paths.

Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>
Change-Id: I23789acc6879086f356ca9accfe400122fd4f0f3
Reviewed-on: http://git.am.freescale.net:8181/32879
Reviewed-by: Marian Cristian Rotariu <marian.rotariu@freescale.com>
[Yang: Original patch taken from
QorIQ-SDK-V1.8-SOURCE-20150619-yocto.iso]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>