crypto: caampkc - workaround for erratum A-006899
The problem will occur only in T1040R1. A workaround
has been verified; set a Scatter Gather (SG) flag
on the Q or G Pointers on KPG, which will give
the PKHA unload time to complete. It's sufficient
to have the SG flag set for only one of these.

Change-Id: I5a403576c6e995110227872c84de7e74da0144c5
Signed-off-by: Tudor Ambarus <tudor.ambarus@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/36617
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-by: Honghua Yin <Hong-Hua.Yin@freescale.com>
[Yang: Original patch taken from
QorIQ-SDK-V1.8-SOURCE-20150619-yocto.iso]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>