rtc: pcf8563 fix: return -EINVAL if we read an invalid time.
commit 538330ccb93849530f5617d1fa870237496caa60 upstream

Return -EINVAL if the voltage low bit is set to avoid getting a bogus
time at boot.
There was a comment stating that util-linux hwclock refuses to set a
new time if we return an error code on read, but at least the current
version do set the time as expected. Remove the comment and the check
for valid time, and let the rtc core check it for us.

Signed-off-by: Jan Kardell <jan.kardell@telliq.com>
Signed-off-by: Alexandre Belloni <alexandre.belloni@free-electrons.com>
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>