powerpc/85xx: Use KCFLAGS when looking for libgcc
Freescale SDK kernel v1.6 and above adds code with dependency on libgcc
to support Freescale DPAA Q/Bman, but the makefile change has not
considered that a cross-compiler normally needs an sysroot option to
correctly report the path of libgcc.

Since CC variable is composed from CROSS_COMPILE variable in top-most
makefile, the sysroot has to be set using KCFLAGS, which is designed to
provide additional options to the C compiler. Moreover, KCFLAGS is not
appended to KBUILD_CFLAGS propagated to architecture specific makefile.
Therefore, this patch modifies the compiler command for querying libgcc
path to use KCFLAGS option, so that once --sysroot=<path> option is set
in KCFLAGS, libgcc can be found correctly.

Signed-off-by: Yuanjie Huang <Yuanjie.Huang@windriver.com>