powerpc/e6500: hardware tablewalk bug fix on per-core spinlock
There is a per-core lock for TLB handlers but been mis-used thus
cause the hardware tablewalk failed.

The root cause of this issue comes from a mis-use of a pointer. That pointer
was used to point to a data structure. And this data structure was 4bytes aligned
so the lowest two bits of its address is always zero. That means the lowest two bits
of this point should always been zero. While in fact, these two bits are not
wasted. It was use as a lock for two threads of the same hard core. 1 is locked and
0 is un-locked. While when using it, some pieces of process ignored it. The pointer
was used directly to get the data structure and data in it. It is right when the lock
was not set but when the lock was set, it is not the address of the data structure.
Such issue will only effect on hardware tablewalk, and will overwrite an very very
important TLB entry: entry 0 of TLB1. This will cause the kernel hangs for ever.

Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>