powerpc: create a function descriptor for init_core_book3e
because of upstream commit 6a3bab90cf78
("powerpc: Remove some unnecessary uses of _GLOBAL() and _STATIC()"),
init_core_book3e was changed to be a local label. But we reference this function
in arch/powerpc/kernel/head_64.S when CONFIG_RELOCATABLE is set. So change it
back to be a function descriptor, or else it will occurr a compile error as below:

 arch/powerpc/kernel/head_64.o: In function `__after_prom_start':
 arch/powerpc/kernel/head_64.S:508: undefined reference to `.init_core_book3e'

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>