powerpc: e6500: set the correct tcd_ptr when boot from non-primary thread
For e6500 core which have threads but no tlbsrx, we use a per-core
lock to avoid the conflicts when the two threads are writing to
the tlb. We have embedded the per-core data struct (tcd) in paca,
and make the tcd_ptr in each thread to point to the tcd in the
primary thread. This is good since the primary thread will be
the first thread to boot in general. But for kdump, the kernel
may boot from a non-primary thread. In this case, we should use
the tcd in the paca of the boot cpu instead of the one in the primary
thread since the tcd data in primary thread will not be initialized
in this case and is just full of rubbish.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>