arm: dts: zynq: disable i2c in PL part
There is something seriously wrong if i2c in PL part is enabled, which will
incur no interrupt coming, and the kernel scheduler is dead as well without
any calltraces or oops even if the related debugging option enabled.

And the further debugging is not easy since there is not so much documents
about these PL parts. What's more, the related PL design about this version
of mini-ITX is not available from Xilinx website.

So disable I2C of PL to avoid this sudden and silent dead.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>