net: xilinx: fix the checksum offload for no udp/tcp packets
As the description in 16.2.6 Checksum Offloading of Zynq-7000 All
Programmable SoC, the Soc only offloads the checksum for udp/tcp
packets, not work for icmp packets.

so software should check if the checksum offloading is correct or
not, before setting skb checksum to lp->ip_summed which is hardcoded
to CHECKSUM_UNNECESSARY, since ip checksum offload always is enabled.

Signed-off-by Roy Li <rongqing.li@windriver.com>