net: dpa: explicitly set the carrier on for XGMII
In commit ("net: dpa: call netif_carrier_off() before register netdev"),
we choose to set the carrier off and assume that the phy state machine
will set the carrier on eventually. But this is not true for XGMII
since we don't support to update the phy link status automatically for
it and even don't start the phy state machine. So we have to explicitly
set the carrier on after starting the mac.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>