tpm: Add sysfs interface to show TPM family major version
So far, there is not a sysfs interface for user space code to
check the TPM family major version(1 or 2). So, add a file
named family_major in /sys/class/tpm/tpmX/ to show it.

This idea is from below usage case:
When tss 1.2 and 2.0 are all installed in one rootfs, we need to
judge which one is used via the sys interface family_major file.

Signed-off-by: Meng Li <Meng.Li@windriver.com>